+++ src/client/gui/sendmessage/client_sendmessageui.h
@@ -41,10 +41,33 @@ class mbClientSendMessageUi : public QDialog
     Q_OBJECT
 
 public:
+    struct Strings
+    {
+        const QString function       ;
+        const QString readAdrType    ;
+        const QString readAddress    ;
+        const QString readFormat     ;
+        const QString readCount      ;
+        const QString writeAdrType   ;
+        const QString writeAddress   ;
+        const QString writeFormat    ;
+        const QString writeCount     ;
+        const QString writeData      ;
+        const QString period         ;
+        const QString writeMaskAnd   ;
+        const QString writeMaskOr    ;
+        const QString settings_prefix;
+        Strings();
+        static const Strings &instance();
+    };
+
+public:
     explicit mbClientSendMessageUi(QWidget *parent = nullptr);
     ~mbClientSendMessageUi();
 
 public:
+    MBSETTINGS cachedSettings() const;
+    void setCachedSettings(const MBSETTINGS &settings);
 
 private Q_SLOTS:
     void setProject(mbCoreProject *p);