+++ changelog.md
@@ -38,3 +38,4 @@
 * Add Modbus Scanner
 * Add new Modbus functions: MASK_WRITE_REGISTER, READ_WRITE_MULTIPLE_REGISTERS
 * Improve 'Send Message' dialog window
+* Add all encodings suppoted by Qt framework to the 'String' data format